Adapter linked resource

+++ src/lib/pages/provider-resources/service-edit.component.ts
@@ -4,12 +4,14 @@ import {DatePipe} from '@angular/common';
 import {ServiceFormComponent} from './service-form.component';
 import {AuthenticationService} from '../../services/authentication.service';
 import {Subscription} from 'rxjs';
-import {Service} from '../../domain/eic-model';
+import {Service, ServiceBundle} from '../../domain/eic-model';
 import {ResourceService} from '../../services/resource.service';
 import {ServiceProviderService} from '../../services/service-provider.service';
 import {NavigationService} from "../../services/navigation.service";
 import {RecommendationsService} from "../../services/recommendations.service";
 import {CatalogueService} from "../../services/catalogue.service";
+import {pidHandler} from "../../shared/pid-handler/pid-handler.service";
+import {FormControlService} from "../../../dynamic-catalogue/services/form-control.service";
 
 @Component({
   selector: 'app-service-edit',
@@ -28,8 +30,10 @@ export class ServiceEditComponent extends ServiceFormComponent implements OnInit
               protected catalogueService: CatalogueService,
               protected injector: Injector,
               public datePipe: DatePipe,
-              public navigationService: NavigationService) {
-    super(injector, authenticationService, serviceProviderService, recommendationsService, catalogueService, route);
+              public navigator: NavigationService,
+              public pidHandler: pidHandler,
+              public dynamicFormService: FormControlService) {
+    super(injector, authenticationService, serviceProviderService, recommendationsService, catalogueService, route, pidHandler, dynamicFormService);
     this.editMode = true;
   }
 
@@ -43,22 +47,75 @@ export class ServiceEditComponent extends ServiceFormComponent implements OnInit
       sessionStorage.removeItem('service');
     } else {
       this.sub = this.route.params.subscribe(params => {
-        this.serviceId = params['resourceId'];
+        // this.serviceId = params['resourceId'];
+        this.serviceId = this.route.snapshot.paramMap.get('resourceId');
         const pathName = window.location.pathname;
         if (pathName.includes('draft-resource/update')) this.pendingService = true;
-        // this.resourceService.getService(this.serviceID).subscribe(service => {
-        this.resourceService[this.pendingService ? 'getPendingService' : 'getServiceBundleById'](this.serviceId, this.catalogueId)
+        if (this.pendingService) {
+          this.resourceService.getPendingService(this.serviceId)
+            .subscribe(service => {
+                this.payloadAnswer = {'answer': {Service: service}};
+                //clear below
+                if (service.mainContact === null) //in case of unauthorized access backend will not show sensitive info
+                  this.navigator.go('/forbidden')
+                ResourceService.removeNulls(service);
+                //remove catalogueId. prefix for same catalogue entries
+                if (service.requiredResources) {
+                  service.requiredResources = service.requiredResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length + 1) : value);
+                }
+                if (service.relatedResources) {
+                  service.relatedResources = service.relatedResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length + 1) : value);
+                }
+                this.formPrepare(service);
+                this.serviceForm.patchValue(service);
+                for (const i in this.serviceForm.controls) {
+                  if (this.serviceForm.controls[i].value === null) {
+                    this.serviceForm.controls[i].setValue('');
+                  }
+                }
+                if (this.serviceForm.get('lastUpdate').value) {
+                  const lastUpdate = new Date(this.serviceForm.get('lastUpdate').value);
+                  this.serviceForm.get('lastUpdate').setValue(this.datePipe.transform(lastUpdate, 'yyyy-MM-dd'));
+                }
+              },
+              err => this.errorMessage = 'Could not get the data for the requested service. ' + err.error,
+              () => {
+                if (window.location.href.indexOf('/add/use-template') > -1) {
+                  this.editMode = false;
+                  this.serviceForm.get('id').setValue('');
+                  this.serviceForm.get('name').setValue('');
+                }
+                if (this.disable) {
+                  this.serviceForm.disable();
+                  this.serviceName = this.serviceForm.get('name').value;
+                } else {
+                  // this.initServiceBitSets();
+                }
+              }
+            );
+        }
+      });
+      if (!this.pendingService) {
+        this.resourceService.getServiceBundleById(this.serviceId, this.catalogueId)
           .subscribe(serviceBundle => {
-              if (serviceBundle.service.mainContact === null) //in case of unauthorized access backend will not show sensitive info
-                this.navigationService.go('/forbidden')
+              this.service = serviceBundle.service;
+              const parsedService = {
+                ...this.service,
+                horizontalService: typeof this.service.horizontalService === 'boolean' ? this.service.horizontalService.toString() : this.service.horizontalService
+              };
+              this.payloadAnswer = {'answer': {Service: parsedService}};
+              //clear below
+
+/*              if (serviceBundle.service.mainContact === null) //in case of unauthorized access backend will not show sensitive info
+                this.navigator.go('/forbidden')
               ResourceService.removeNulls(serviceBundle.service);
               //remove catalogueId. prefix for same catalogue entries
               if (serviceBundle.service.requiredResources) {
-                serviceBundle.service.requiredResources = serviceBundle.service.requiredResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length+1) : value);
+                serviceBundle.service.requiredResources = serviceBundle.service.requiredResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length + 1) : value);
               }
               if (serviceBundle.service.relatedResources) {
-                serviceBundle.service.relatedResources = serviceBundle.service.relatedResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length+1) : value);
-              }
+                serviceBundle.service.relatedResources = serviceBundle.service.relatedResources.map(value => value.startsWith(this.catalogueId) ? value.substring(this.catalogueId.length + 1) : value);
+              }*/
               this.formPrepare(serviceBundle.service);
               this.serviceForm.patchValue(serviceBundle.service);
               for (const i in this.serviceForm.controls) {
@@ -82,11 +139,11 @@ export class ServiceEditComponent extends ServiceFormComponent implements OnInit
                 this.serviceForm.disable();
                 this.serviceName = this.serviceForm.get('name').value;
               } else {
-                this.initServiceBitSets();
+                // this.initServiceBitSets();
               }
             }
           );
-      });
+      }
     }
   }
 
@@ -94,7 +151,7 @@ export class ServiceEditComponent extends ServiceFormComponent implements OnInit
     super.onSubmit(service, tempSave, this.pendingService);
   }
 
-  initServiceBitSets() {
+  /*initServiceBitSets() {
     this.handleBitSets(0, 0, 'name');
     this.handleBitSets(0, 21, 'abbreviation');
     this.handleBitSets(0, 1, 'resourceOrganisation');
@@ -117,6 +174,6 @@ export class ServiceEditComponent extends ServiceFormComponent implements OnInit
     this.handleBitSets(9, 22, 'termsOfUse');
     this.handleBitSets(9, 23, 'privacyPolicy');
     this.handleBitSets(10, 19, 'orderType');
-  }
+  }*/
 
 }
